Releasy and Swappy Update
Yes. The time has come. this should be ready for release. In this update i finished the Options Menu, adding external link to this repo, ability to change the background of the game and of the ball, made EVERY sound load before startup, (this has increased the startup duration tho) fixed a few bugs and other stuff. Added also the possibility of changing BGM. Or fixed. Can't remember if i've committed that before or not lol. Also changing the volume now doesn't restart the song.  Oh i've changed to a new Icon and Splash too!